+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### 9.3.4 SNAPSHOT
   * (GH-63) Added message type versions in categories: acmt, admi, auth, caaa, camt, catm, fxtr, pacs, reda, seev, semt, sese, setr, supl and trck
   * Added new business process and messages for: caad, cafc, cain, casp
+  * Removed obsolete trea (treasury) message types
   * Added model and support for the BusinessApplicationHeaderV03
 
 #### 9.3.3 - October 2022